variable header verb, cleanup
