+++ web-common/src/layout/navigation/Navigation.svelte
@@ -36,6 +36,8 @@
 
   $: isModelerEnabled = $readOnly === false;
 
+  $: console.log("RUNTIME", $runtime);
+
   $: instance = createRuntimeServiceGetInstance($runtime.instanceId);
   $: olapConnector = $instance.data?.instance?.olapConnector;
   $: isModelingSupportedForCurrentOlapDriver =
